+++ tests/testthat/test-estimate_truncation.R
@@ -14,9 +14,11 @@ test_that("estimate_truncation can return values from simulated data and plot
   )
   expect_equal(
     names(est),
-    c("dist", "obs", "last_obs", "cmf", "data", "fit")
+    c("fit", "args", "observations", ".internal")
   )
-  expect_s3_class(est$dist, "dist_spec")
+  expect_s3_class(est, "estimate_truncation")
+  expect_s3_class(est, "epinowfit")
+  expect_s3_class(summary(est), "dist_spec")
   expect_error(plot(est), NA)
 })
 
@@ -32,9 +34,11 @@ test_that("estimate_truncation can return values from simulated data with the
   )))
   expect_equal(
     names(est),
-    c("dist", "obs", "last_obs", "cmf", "data", "fit")
+    c("fit", "args", "observations", ".internal")
   )
-  expect_s3_class(est$dist, "dist_spec")
+  expect_s3_class(est, "estimate_truncation")
+  expect_s3_class(est, "epinowfit")
+  expect_s3_class(summary(est), "dist_spec")
   expect_error(plot(est), NA)
 })
 
@@ -58,16 +62,16 @@ test_that("estimate_truncation works with filter_leading_zeros set", {
   )
   expect_named(
     modified_data_fit,
-    c("dist", "obs", "last_obs", "cmf", "data", "fit")
+    c("fit", "args", "observations", ".internal")
   )
   # Compare the results of the two fits
   expect_equal(
-    original_data_fit$dist$dist,
-    modified_data_fit$dist$dist
+    summary(original_data_fit)$dist,
+    summary(modified_data_fit)$dist
   )
   expect_equal(
-    original_data_fit$data$obs_dist,
-    modified_data_fit$data$obs_dist
+    original_data_fit$args$obs_dist,
+    modified_data_fit$args$obs_dist
   )
 })
 
@@ -81,8 +85,10 @@ test_that("estimate_truncation works with zero_threshold set", {
   out <- estimate_truncation(modified_data,
     verbose = FALSE, chains = 2, iter = 1000, warmup = 250
   )
-  expect_named(out, c("dist", "obs", "last_obs", "cmf", "data", "fit"))
-  expect_s3_class(out$dist, "dist_spec")
+  expect_named(out, c("fit", "args", "observations", ".internal"))
+  expect_s3_class(out, "estimate_truncation")
+  expect_s3_class(out, "epinowfit")
+  expect_s3_class(summary(out), "dist_spec")
 })
 
 options(old_opts)